metric_calc function edits to df output.
The edit adds columns for run_number and variable name (which will probably be changed later).

Ordered the resulting df to be run_number_varible, then metric value. can delete if not needed but I think it is cleaner.t.

The edit adds columns for run_number and variable name (which will probably be changed later).

Ordered the resulting df to be run_number_varible, then metric value. can delete if not needed but I think it is cleaner.