Editing prob_calc and adding hector_result ex data
This commit includes edits to prob_cal to make it simpler by replacing the use of dfs with a vector of values.

This also includes the addition of hector_result  which will serve as an example data frame for the package and can be used in examples and test code.

The last major change has been the removal of an error code in the metric_calc that was preventing the ability to run metric_calc on a hector result that has more than one variable in the output.