Vignette edits - re-organization for new_metric
Moved new_metric up into the set-up under the header "Defining output metric"

Moved portion that defines a scoring criterion to  the iterate_hector header. Presents the criterion concept after mentioning that it is needed for running iterate_hector.

Edited the text on scoring hector to try and simplify the score_ramp explanation.