Creating function and documentation for joint_pdf_sample
This function is on the larger side and can likely broken into multiple function if that will make it easier to test/ if it will be more robust.

Additionally, this function is written specifically for ECS and diffusivity. It will be beneficial to make this suitable for accepting any variables for which we may want to build joint PDFs